requirements.txt
Django~=4.1.7
djangorestframework==3.14.0
psycopg2==2.9.5
six==1.16.0